refactor: Pass the remind invoked a notification to remind method.
In the future, there will be not only one remind objects
belong to a task, so method service.task.remind must not retrive the
current remind object from within a task, it must be passed from
outside of the method. Also, the API for reminding a task must be
removed, due to it no longer suits the signature of service.task.remind.