feat: Add buttons for switching between different views.
The first view is calendar view, it shows the tasks scheduled today at
the remaining time. This is the default view. The second view is the
following tasks view, it shows all the following tasks on a single
page; The third one is the tasks view, shows all tasks by a pagination
style, from latest to oldest.